DRT-5178 Break manifests source into separate provider for test seam
