+++ server/src/main/scala/services/graphstages/ArrivalsGraphStage.scala
@@ -15,8 +15,11 @@ import scala.language.postfixOps
 
 
 sealed trait ArrivalsSourceType
+
 case object LiveArrivals extends ArrivalsSourceType
+
 case object ForecastArrivals extends ArrivalsSourceType
+
 case object BaseArrivals extends ArrivalsSourceType
 
 class ArrivalsGraphStage(name: String = "",
@@ -98,17 +101,9 @@ class ArrivalsGraphStage(name: String = "",
     }
 
     def mergeUpdatesAndPurge(updates: Map[Int, Arrival], existingArrivals: Map[Int, Arrival]): Map[Int, Arrival] = {
-      val updated = updates
-        .foldLeft(existingArrivals) {
-          case (soFar, (newArrivalId, newArrival)) => soFar.updated(newArrivalId, newArrival)
-        }
-
-      val minusExpired = purgeExpired(updated)
-
-      val numPurged = existingArrivals.size - minusExpired.size
-      if (numPurged > 0) log.info(s"Purged $numPurged expired arrivals during update")
-
-      minusExpired
+      updates.foldLeft(existingArrivals) {
+        case (soFar, (newArrivalId, newArrival)) => soFar.updated(newArrivalId, newArrival)
+      }
     }
 
     def purgeExpired(flightsById: Map[Int, Arrival]): Map[Int, Arrival] = {
@@ -116,11 +111,6 @@ class ArrivalsGraphStage(name: String = "",
       flightsById.filterNot {
         case (_, arrival) => expired(arrival)
       }
-    }
-
-    def purgeExpired(flights: Set[Arrival]): Set[Arrival] = {
-      val expired: Arrival => Boolean = Crunch.hasExpired(now(), expireAfterMillis, (a: Arrival) => a.PcpTime.getOrElse(0L))
-      flights.filterNot(expired)
     }
 
     def mergeAllSourcesAndPush(baseArrivals: Map[Int, Arrival], forecastArrivals: Map[Int, Arrival], liveArrivals: Map[Int, Arrival]): Unit = {
@@ -188,7 +178,7 @@ class ArrivalsGraphStage(name: String = "",
             case None =>
               (notFoundSoFar + 1, mergedSoFar)
             case Some(baseArrival) =>
-              val actPax = forecastArrival.ActPax.filter(_>0).orElse(baseArrival.ActPax)
+              val actPax = forecastArrival.ActPax.filter(_ > 0).orElse(baseArrival.ActPax)
               val mergedArrival = baseArrival.copy(ActPax = actPax, TranPax = forecastArrival.TranPax, Status = forecastArrival.Status)
               (notFoundSoFar, mergedSoFar.updated(fcstId, mergedArrival))
           }
@@ -202,8 +192,8 @@ class ArrivalsGraphStage(name: String = "",
           val mergedArrival = liveArrival.copy(
             rawIATA = baseArrival.rawIATA,
             rawICAO = baseArrival.rawICAO,
-            ActPax = liveArrival.ActPax.filter(_>0).orElse(mergedSoFarArrival.ActPax),
-            TranPax = liveArrival.ActPax.filter(_ > 0).flatMap(actPax=> liveArrival.TranPax).orElse(mergedSoFarArrival.TranPax))
+            ActPax = liveArrival.ActPax.filter(_ > 0).orElse(mergedSoFarArrival.ActPax),
+            TranPax = liveArrival.ActPax.filter(_ > 0).flatMap(_ => liveArrival.TranPax).orElse(mergedSoFarArrival.TranPax))
 
           mergedSoFar.updated(liveId, mergedArrival)
       }
