+++ server/src/main/scala/services/graphstages/VoyageManifestsGraphStage.scala
@@ -1,17 +1,9 @@
 package services.graphstages
 
-import java.io.InputStream
-import java.nio.charset.StandardCharsets.UTF_8
-import java.util.zip.ZipInputStream
-
-import akka.NotUsed
 import akka.actor.ActorSystem
-import akka.stream.scaladsl.{Sink, Source, StreamConverters}
-import akka.stream.stage.{GraphStage, GraphStageLogic, OutHandler}
 import akka.stream._
-import akka.util.ByteString
-import com.amazonaws.auth.profile.ProfileCredentialsProvider
-import com.mfglabs.commons.aws.s3.{AmazonS3AsyncClient, S3StreamBuilder}
+import akka.stream.stage.{GraphStage, GraphStageLogic, OutHandler}
+import drt.server.feeds.api.S3ApiProvider
 import drt.shared.CrunchApi.MillisSinceEpoch
 import drt.shared.DqEventCodes
 import org.slf4j.{Logger, LoggerFactory}
@@ -20,11 +12,9 @@ import passengersplits.parsing.VoyageManifestParser.VoyageManifest
 import server.feeds.{ManifestsFeedFailure, ManifestsFeedResponse, ManifestsFeedSuccess}
 import services.SDate
 
-import scala.collection.immutable.Seq
-import scala.collection.mutable.ArrayBuffer
+import scala.concurrent.Await
 import scala.concurrent.ExecutionContext.Implicits.global
 import scala.concurrent.duration._
-import scala.concurrent.{Await, Future}
 import scala.language.postfixOps
 import scala.util.matching.Regex
 import scala.util.{Failure, Success, Try}
@@ -42,8 +32,8 @@ case class DqManifests(lastSeenFileName: String, manifests: Set[VoyageManifest])
   }
 }
 
-class VoyageManifestsGraphStage(bucketName: String,
-                                portCode: String,
+class VoyageManifestsGraphStage(portCode: String,
+                                provider: S3ApiProvider,
                                 initialLastSeenFileName: String,
                                 minCheckIntervalMillis: MillisSinceEpoch = 30000)
                                (implicit actorSystem: ActorSystem,
@@ -98,15 +88,16 @@ class VoyageManifestsGraphStage(bucketName: String,
 
   def fetchNewManifests(startingFileName: String): ManifestsFeedResponse = {
     log.info(s"Fetching manifests from files newer than $startingFileName")
-    val eventualFileNameAndManifests = manifestsFuture(startingFileName)
+    val eventualFileNameAndManifests = provider
+      .manifestsFuture(startingFileName)
       .map(fetchedFilesAndManifests => {
-        val (latestFileName, fetchedManifests) = if (fetchedFilesAndManifests.isEmpty) {
-          (startingFileName, Set[Option[VoyageManifest]]())
-        } else {
+        val (latestFileName, fetchedManifests) = if (fetchedFilesAndManifests.nonEmpty) {
           val lastSeen = fetchedFilesAndManifests.map { case (fileName, _) => fileName }.max
-          val manifests = fetchedFilesAndManifests.map { case (_, manifest) => manifest }.toSet
+          val manifests = fetchedFilesAndManifests.map { case (_, manifest) => jsonStringToManifest(manifest) }.toSet
           (lastSeen, manifests)
         }
+        else (startingFileName, Set[Option[VoyageManifest]]())
+
         (latestFileName, fetchedManifests)
       })
 
@@ -123,78 +114,17 @@ class VoyageManifestsGraphStage(bucketName: String,
     }
   }
 
-  def manifestsFuture(latestFile: String): Future[Seq[(String, Option[VoyageManifest])]] = {
-    log.info(s"Requesting DQ zip files > ${latestFile.take(20)}")
-    zipFiles(latestFile)
-      .mapAsync(64) { filename =>
-        log.info(s"Fetching $filename")
-        val zipByteStream = S3StreamBuilder(s3Client).getFileAsStream(bucketName, filename)
-        Future(fileNameAndContentFromZip(filename, zipByteStream, portCode, None))
-      }
-      .mapConcat(identity)
-      .runWith(Sink.seq[(String, Option[VoyageManifest])])
-  }
-
-  def zipFiles(latestFile: String): Source[String, NotUsed] = {
-    filterToFilesNewerThan(filesAsSource, latestFile)
-  }
-
-  def fileNameAndContentFromZip[X](zipFileName: String,
-                                   zippedFileByteStream: Source[ByteString, X],
-                                   port: String,
-                                   maybeAirlines: Option[List[String]]): List[(String, Option[VoyageManifest])] = {
-    val inputStream: InputStream = zippedFileByteStream.runWith(
-      StreamConverters.asInputStream()
-    )
-    val zipInputStream = new ZipInputStream(inputStream)
-    val zipFilesAndMaybeManifests = Stream
-      .continually(zipInputStream.getNextEntry)
-      .takeWhile(_ != null)
-      .map { _ =>
-        val buffer = new Array[Byte](4096)
-        val stringBuffer = new ArrayBuffer[Byte]()
-        var len: Int = zipInputStream.read(buffer)
-
-        while (len > 0) {
-          stringBuffer ++= buffer.take(len)
-          len = zipInputStream.read(buffer)
+  def jsonStringToManifest(content: String): Option[VoyageManifest] = {
+    VoyageManifestParser.parseVoyagePassengerInfo(content) match {
+      case Success(m) =>
+        if (m.EventCode == DqEventCodes.DepartureConfirmed && m.ArrivalPortCode == portCode) {
+          log.info(s"Using ${m.EventCode} manifest for ${m.ArrivalPortCode} arrival ${m.flightCode}")
+          Option(m)
         }
-        val content: String = new String(stringBuffer.toArray, UTF_8)
-        VoyageManifestParser.parseVoyagePassengerInfo(content) match {
-          case Success(m) =>
-            if (m.EventCode == DqEventCodes.DepartureConfirmed && m.ArrivalPortCode == port) {
-              log.info(s"Using ${m.EventCode} manifest for ${m.ArrivalPortCode} arrival ${m.flightCode}")
-              (zipFileName, Option(m))
-            } else (zipFileName, None)
-        }
-      }
-      .toList
-
-    val relevantCount = zipFilesAndMaybeManifests.count { case (_, maybeVm) => maybeVm.isDefined }
-    log.info(s"$relevantCount relevant manifests out of ${zipFilesAndMaybeManifests.length} received in $zipFileName")
-
-    zipFilesAndMaybeManifests
-  }
-
-  def filterToFilesNewerThan(filesSource: Source[String, NotUsed], latestFile: String): Source[String, NotUsed] = {
-    val filterFrom: String = filterFromFileName(latestFile)
-    filesSource.filter(fn => fn >= filterFrom && fn != latestFile)
-  }
-
-  def filterFromFileName(latestFile: String): String = {
-    latestFile match {
-      case dqRegex(dateTime, _) => dateTime
-      case _ => latestFile
+        else None
+      case Failure(t) =>
+        log.error(s"Failed to parse voyage manifest json", t)
+        None
     }
   }
-
-  def filesAsSource: Source[String, NotUsed] = {
-    S3StreamBuilder(s3Client)
-      .listFilesAsStream(bucketName)
-      .map {
-        case (filename, _) => filename
-      }
-  }
-
-  def s3Client: AmazonS3AsyncClient = new AmazonS3AsyncClient(new ProfileCredentialsProvider("drt-prod-s3"))
 }
