+++ server/src/test/scala/services/crunch/ForecastCrunchSpec.scala
@@ -208,7 +208,7 @@ class ForecastCrunchSpec extends CrunchTestLike {
     val crunch = runCrunchGraph(now = () => SDate(forecastScheduled).addDays(-1))
 
     offerAndWait(crunch.forecastArrivalsInput, ArrivalsFeedSuccess(forecastArrivals))
-    crunch.forecastTestProbe.expectNoMsg(1 seconds)
+    crunch.forecastTestProbe.expectNoMessage(1 seconds)
 
     true
   }
